+++ test/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 
+use TxTextControl\ReportingCloud\Console\Helper;
 use TxTextControl\ReportingCloud\Exception\RuntimeException;
-use TxTextControl\ReportingCloud\CliHelper as Helper;
 
 $filenameAutoload = realpath(__DIR__ . '/../vendor/autoload.php');
 
@@ -13,9 +13,7 @@ include_once $filenameAutoload;
 
 // ---------------------------------------------------------------------------------------------------------------------
 
-if (null === Helper::username() || null === Helper::password()) {
-    echo Helper::errorMessage();
-    exit(1);
-}
+Helper::checkCredentials();
 
 // ---------------------------------------------------------------------------------------------------------------------
+