+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 
 ## 1.0.10 - 2016-07-20
 
-* Moved orphaned helper functions (`reporting_cloud_*`) to helper class `\TxTextControl\ReportingCloud\CliHelper`.
+* Moved orphaned helper functions (`reporting_cloud_*`) to helper class `\TxTextControl\ReportingCloud\Console\CliHelper`.
 
 ## 1.0.9 - 2016-07-18